Crate async_unlock

This crate describes functions to work with async_std’s Mutex in a manner that guarantees shortest possible lock time by dropping MutexGuard ASAP.

Functions

unlock

Performs computation on a value inside Mutex, unlocking Mutex immediately after f is computed

unlock_f

Performs async computation on a value inside Mutex This function guarantees that Mutex will be unlocked before awaiting computation described by f